app/app/requirements.txt
# utils
pytest~=6.2.4
python-dotenv~=0.18.0
requests~=2.25.1
typing-extensions~=3.10.0.0

# web app
dash~=1.20.0
plotly==5.0.0
dash-bootstrap-components~=0.12.2
dash-bootstrap-templates~=0.1.1
dash-daq~=0.5.0
dash-cytoscape~=0.3.0
uwsgi~=2.0.19.1

# modelling
networkx~=2.5.1
pandas~=1.1.5
